fix: converted wait interval values to long to avoid casting a long (timeout) to an int

+++ src/main/java/com/synopsys/integration/wait/ResilientJobConfig.java
@@ -41,7 +41,7 @@ public class ResilientJobConfig {
         return startTimeSupplier.get();
     }
 
-    public int getWaitIntervalInSeconds() {
+    public long getWaitIntervalInSeconds() {
         return waitIntervalTracker.getNextWaitIntervalInSeconds();
     }
 }
