+++ src/main/java/com/synopsys/integration/wait/tracker/WaitIntervalTracker.java
@@ -10,5 +10,5 @@ package com.synopsys.integration.wait.tracker;
 public interface WaitIntervalTracker {
 
     long getTimeoutInSeconds();
-    int getNextWaitIntervalInSeconds();
+    long getNextWaitIntervalInSeconds();
 }
